wakelni-backend/requirements.txt

django>=5.0
djangorestframework
djangorestframework-simplejwt
psycopg2-binary
